LPS-95267 Move to DTO package
