+++ modules/apps/data-engine/data-engine-spi/src/main/java/com/liferay/data/engine/spi/dto/SPIDataDefinitionField.java
@@ -12,7 +12,7 @@
  * details.
  */
 
-package com.liferay.data.engine.spi.definition;
+package com.liferay.data.engine.spi.dto;
 
 import com.liferay.petra.lang.HashUtil;
