+++ modules/apps/data-engine/data-engine-spi/src/main/java/com/liferay/data/engine/spi/field/type/FieldType.java
@@ -14,7 +14,7 @@
 
 package com.liferay.data.engine.spi.field.type;
 
-import com.liferay.data.engine.spi.definition.SPIDataDefinitionField;
+import com.liferay.data.engine.spi.dto.SPIDataDefinitionField;
 import com.liferay.portal.kernel.json.JSONObject;
 
 import java.util.Map;
